updated mysql properties for docker
The mysql.properties that gets built into the tracking services war is not modifable at run time. Try to override the values with environment varaibles failed because they are good only for the original process not the processes spawned off of it. Therefore, updated two files to break the hardcoded host name, ports, user names, password, and even database type in the end.

In the Dockerfile.* copied the mysql.properties directly to /etc/tracking_service/jdbc.properties. This file, tracking_service/jdbc.properties, can then be replaced with the docker run --volume or --mount options. This overrides all other options in looking up these parameters.

Updated DBConnector.java to look at the /etc file first, then sytem properties, then default properties in the WAR. The change leaves the current installs and operation as is.